Add optional argument `sack` to HelloFrame to make it possible to send a SACK before the server sends you strings that you may have already received.
The protocol used to make it possible to SACK before subscribing to boxes,
but on a 2010-04-17 I made a slightly misguided commit
"Remove the gimme_boxes frame type; move this boolean (and the
succeedsTransport argument) into the hello frame.", which intended to
reduce the amount of state that ServerTransport needs to keep track
of. One of the motivations for the commit was that I did not want to
keep track of whether gimme_boxes was already sent, and send some kind
of TransportKillFrame if client sent it twice. gimme_boxes existed in the
first place was so that user could send:

hello_frame
sack
gimme_boxes

After the 2010-04-17 change, gimme_boxes(succeedsTransport=...) became part of
HelloFrame, but it was not possible to SACK in the HelloFrame. This was
a mistake. Clients need to be able to SACK before the server writes
strings over the transports.

One approach would have been to resurrect `gimme_boxes` as GimmeStringsFrame,
but I decided to take the other apporach of putting more arguments into
HelloFrame. Now client sends something like this:

HelloFrame(succeedsTransport=null [implies become primary], sack=SACK(...))

That passes the SACK to Stream before it writes boxes over the transport.
Client may still send SackFrames at any time.

This commit fixes the client as well:
client.js now uses this HelloFrame(sack=) functionality so that it
doesn't get the same strings twice.