Implement cw.net.XDRTracker, which can dynamically create iframes for cross-domain XHR.  This can create as many iframes as necessary for the active Streams.  XDRTracker obsoletes bootstrap_XDRSetup.js, which always created two iframes at page load time.
Implement an easier-to-use HTTP endpoint NiceHttpEndpoint which lets you
use a %random% token for a random subdomain, and doesn't require you to
specify a secondaryUrl.

Rename HttpEndpoint to ExpandedHttpEndpoint_.

Remove the "redirect if wrong location" code that worked around a Firefox 3
iframe bug; we might add it back later.  We'll soon implement CORS anyway,
and that will work in Firefox 3.5+.